requirements.txt
asgiref==3.8.1
Django==4.2.10
djangorestframework==3.15.0
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
sqlparse==0.5.0
